fold LighthouseClaimsOverview into ClaimsOverview
