+++ modules/mobile/app/controllers/mobile/v0/claims_and_appeals_controller.rb
@@ -90,7 +90,7 @@ module Mobile
 
       # It was found that FE is using both different casing between multi image upload and single image upload.
       # This shouldn't matter due to the x-key-inflection: camel header being used but that header only works if the
-      # body payload is in json, which the single doc upload is not (at least in specs for both LH and EVSS).
+      # body payload is in json, which the single doc upload is not (at least in the specs).
       def tracked_item_id
         params[:trackedItemId] || params[:tracked_item_id]
       end
