+++ modules/mobile/app/services/mobile/v0/lighthouse_claims/claims_index_interface.rb
@@ -73,7 +73,7 @@ module Mobile
         end
 
         def claims_adapter
-          Mobile::V0::Adapters::LighthouseClaimsOverview.new
+          Mobile::V0::Adapters::ClaimsOverview.new
         end
       end
     end
